+++ tests/sdks/dotnet/Infrastructure/SecretProviderFactoryTests.cs
@@ -20,7 +20,7 @@ public class SecretProviderFactoryTests : IDisposable
 			Environment.SetEnvironmentVariable(name, value);
 		}
 
-		if (_tempDirToDelete is not null)
+		if (_tempDirToDelete is not null && Directory.Exists(_tempDirToDelete))
 		{
 			Directory.Delete(_tempDirToDelete, true);
 		}
